api/requirements.txt
click==7.1.2
cycler==0.10.0
Flask==1.1.2
Flask-Cors==3.0.9
itsdangerous==1.1.0
Jinja2==2.11.2
kiwisolver==1.3.1
MarkupSafe==1.1.1
matplotlib==3.3.3
numpy==1.19.4
Pillow==8.0.1
pip==19.0.3
pyparsing==2.4.7
python-dateutil==2.8.1
setuptools==40.8.0
six==1.15.0
Werkzeug==1.0.1